refactor: rename service worker to use lower case for imports
