+++ src/providers/service-worker.js
@@ -1,4 +1,4 @@
-import * as serviceWorker from '../serviceWorker';
+import * as serviceWorker from '../sw';
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
